v1.5.0 - BACnet

+++ LightController/ErrorBox.cs
@@ -1,20 +1,18 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Diagnostics;
 using System.Windows;
 
 namespace LightController
 {
     public static class ErrorBox
     {
-        public static void Show(string msg)
+        public static void Show(string msg, bool kill = true)
         {
             MessageBox.Show(msg, "Light Controller", MessageBoxButton.OK, MessageBoxImage.Error, MessageBoxResult.None, MessageBoxOptions.DefaultDesktopOnly);
-            LogFile.Error("Application closed: '" + msg + "'");
-            Process.GetCurrentProcess().Kill();
+            if(kill)
+            {
+                LogFile.Error("Application closed: '" + msg + "'");
+                Process.GetCurrentProcess().Kill();
+            }
         }
 
         /// <summary>
